main_automation_script.py
"""
main_automation_script.py - Main Script for Py_Automation

This script orchestrates the automation process by integrating various modules.

Author: [Your Name]
"""

from modules.process_identification import identify_repetitive_processes
from modules.modular_development import modular_script_a, modular_script_b
from modules.api_integration import integrate_with_api
from modules.error_reduction import reduce_errors
from utils.constants import PROCESS_A_THRESHOLD, API_BASE_URL
from utils.helper_functions import validate_input

def main():
    # Identify repetitive processes
    identified_processes = identify_repetitive_processes()
    print(f"Identified Repetitive Processes: {identified_processes}")

    # Perform modular development
    result_a = modular_script_a()
    result_b = modular_script_b()
    print(f"Modular Script A Result: {result_a}")
    print(f"Modular Script B Result: {result_b}")

    # Integration with API
    api_data = {"key": "value"}
    api_response = integrate_with_api(API_BASE_URL, api_data)
    print(f"API Integration Response: {api_response}")

    # Error reduction measures
    input_data = {"required_field": "value"}
    if validate_input(input_data):
        error_reduction_result = reduce_errors()
        print(f"Error Reduction Result: {error_reduction_result}")
    else:
        print("Input validation failed.")

if __name__ == "__main__":
    main()
