tests/test_main_automation_script.py
"""
test_main_automation_script.py - Tests for Main Automation Script

This module contains tests for the main automation script.

Author: [Your Name]
"""

import unittest
from unittest.mock import patch, MagicMock
from main_automation_script import main

class TestMainAutomationScript(unittest.TestCase):
    @patch('modules.process_identification.identify_repetitive_processes', return_value=["Process A"])
    @patch('modules.modular_development.modular_script_a', return_value="Script A Result")
    @patch('modules.modular_development.modular_script_b', return_value="Script B Result")
    @patch('modules.api_integration.integrate_with_api', return_value="API Integration Result")
    @patch('modules.error_reduction.reduce_errors', return_value="Error Reduction Result")
    @patch('utils.helper_functions.validate_input', return_value=True)
    def test_main_script_execution(self, mock_identify, mock_script_a, mock_script_b, mock_api, mock_reduce_errors, mock_validate_input):
        # Test the main script execution
        with patch('builtins.print') as mock_print:
            main()

        # Assert that the expected functions were called
        mock_identify.assert_called_once()
        mock_script_a.assert_called_once()
        mock_script_b.assert_called_once()
        mock_api.assert_called_once()
        mock_reduce_errors.assert_called_once()
        mock_validate_input.assert_called_once()

        # Assert that the expected outputs were printed
        expected_output = [
            "Identified Repetitive Processes: ['Process A']",
            "Modular Script A Result: Script A Result",
            "Modular Script B Result: Script B Result",
            "API Integration Response: API Integration Result",
            "Error Reduction Result: Error Reduction Result"
        ]
        mock_print.assert_has_calls([unittest.mock.call(output) for output in expected_output])

    @patch('utils.helper_functions.validate_input', return_value=False)
    def test_main_script_input_validation_failure(self, mock_validate_input):
        # Test main script with input validation failure
        with patch('builtins.print') as mock_print:
            main()

        # Assert that input validation failure message was printed
        mock_print.assert_called_once_with("Input validation failed.")

if __name__ == '__main__':
    unittest.main()
